requirements.txt
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.2.0
polars==0.16.16
protobuf==4.22.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
